End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -10,6 +10,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -60,7 +61,7 @@ namespace AuctionService.Controllers
 
             return _mapper.Map<AuctionDTO>(auction);
         }
-
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDTO)
         {
@@ -68,7 +69,7 @@ namespace AuctionService.Controllers
 
             //TODO: add curent user as seller
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -83,6 +84,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, _mapper.Map<AuctionDTO>(auction));
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
         {
@@ -90,7 +92,7 @@ namespace AuctionService.Controllers
 
             if(auction == null) return NotFound();
 
-            //TODO: check seller == username
+            if(auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
@@ -111,6 +113,7 @@ namespace AuctionService.Controllers
 
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -118,15 +121,22 @@ namespace AuctionService.Controllers
 
             if(auction == null) return NotFound();
 
-            //TODO: check seller == username
+            if(auction.Seller != User.Identity.Name) return Forbid();
 
-            await _publishEndpoint.Publish(new {Id = auction.Id.ToString()});
+            try
+            {
+                await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
 
-            _context.Auctions.Remove(auction);
+                _context.Auctions.Remove(auction);
 
-            var result = await _context.SaveChangesAsync() > 0;
+                var result = await _context.SaveChangesAsync() > 0;
 
-            if(!result) return BadRequest("Could not delete auction in the DB");
+                if(!result) return BadRequest("Could not delete auction in the DB");
+            }
+            catch(Exception ex)
+            {
+                System.Console.WriteLine(ex.Message);
+            }
 
             return Ok();
         }
